Update to Bulma v1
Update Bulma list to do away with using Bulma's internal sass variables
in favour of new css variables. Created a matching set of css variables
for Bulma list so people can choose to keep using the old sass ones for
customisation in their project, or switch to the newer css variables (
prefixed with be- for 'bulma extension' to help differentiate them in
the browser devtools as there are a lot of these in Bulma v1!

Updates the documentation slightly to better match the style of Bulma's
updated docs.